requirements.txt
streamlit
fastapi
python-dotenv
google-generativeai
openai-whisper
uvicorn
langchain
langchain_community
langchain_core
moviepy
pymongo[srv]
-e .